Introduce a new build option called BOOST_NO_GLOBAL_FILESYSTEM_NAMESPACE.
Systems like FreeBSD Capsicum and Nuxi CloudABI provide purely
capability-based runtime environments. In these environments it is
possible to access the filesystems only through handles to directories
as part of their security model. There is no global filesystem
namespace.

Most of Boost seems to build pretty well in these environments, with the
exception of classes that operate on file contents. Many of these
classes allow you to pass in file streams, but also provide utility
functions that open these files on your behalf. These will need to be
disabled in this environment.