Update PGI C++ compiler support
When the PGI C++ compiler changed to be GNU compatible, pgi.hpp stopped being used, because the check for __GNUC__ in select_compiler_config.hpp was true before the preprocessor ever got to the check for __PGI.  Rearrange the order of the checks in select_compiler_config.hpp, moving the check for __PGI above the check for __GNUC__.

pgi.hpp was designed for a very old version of PGI C++, before it was GNU compatible.  The settings in that file won't work for PGI compilers of the last few years.  Replace the entire file with one that just includes gcc.hpp and then adjusts a few macros for areas where PGI is not quite GNU compatible.  (The old PGI compilers are not actively being used by any customers that we (PGI) know of, so keeping the old contents of pgi.hpp would be of little or no benefit.)